Wrote up the ruby attack
I wanted to see what the code would look like for the Obj-C and ruby
based attack that we have in different repos. The tests are a little
expanded - just to make sure it's all working, and the times are a little
disappointing, but maybe Bret will see something I missed.