Added repeated testing
I saw that I could have simplified the code a bit - so I did that, and then
added in the multiple testing to make sure that the JIT was all ready to go
and the loading from the words was not an issue. Still... the performance of
Bret's code is far better.