Added the grouping of the words
I added in the grouping of the words based on their length, and that made
a little difference in the speed of execution, but I really need to have
some better logging times on the attack function as I think it's in there
or maybe in one of the others that's simple to write, but costly to execute.
For now, this is still a distant second to Bret's code, but it's getting
there.