Added logging and a main
I wanted to be able to make a little more standard runs with performance
logging, so I pulled in Gary's logging namespace which is just what the
doctor ordered, and then added a main so that we can 'lein run' this guy
and have logging to `log/cq.log`. Just a few pieces to make performance
tuning a little easier.