Added hit counts to maps
I wanted to know how many possibles there were for each cyphertext, and the
simplest way was to compute it when computing the list and then just using it
later in the log message. While I was at it, there's no reason to have two
of the same cypherwords in the attack - so let's make sure they are distinct.
Still this does nothing for the execution time. Very odd.