Getting closer with the speed
The bottleneck was the creation of the list of possible plaintext words for
each cyphertext word. I've refactored it a bit and put in a pmap to spread
the load across the words as this is all about lookups and calls. While I was
at it, I decided to make the possible? function more efficient by checking the
two words for equal length of distinct characters and only if that were true
would we continue on with the next step. The hope is that this would allow
more of a "fail fast" on the function.