Removed the word list from the solver
We don't need to pass in the list of words as they are defined within the
`block` namespace, and so are part of the solution already. This makes the
code look a lot cleaner from the outside, and also they don't have to know
about the organization of the words - as it's now more than a simple list.