Used Bret's idea of non-distinct pattern
If we relax the distinct requirement on the `pattern`, we still get something
that we can use for the matching as it's deterministic and it just has a slight
offset for those words with a pattern of repeated characters, but it all still
works. So with that, we get the call to pattern down another 40%. Very nice.