Added traditional endpoint for solver
I had a little time, and I wanted to add in the endpoint for solving the
general quip - taking the same kind of data that it produces, and then
running it through the solver and sending back the response. I've checked
this out with Postman in Chrome, and it's working like a charm. Very nice.