Went back to keywords in JSON
I really don't like the string keys in JSON for clojure, so I went back to
keywords, but then I had to change the conversion of the clue to use the 'name'
function on the keyword before getting the first character. Not bad, and I like
the way the code looks, and works, better now.