Cleaned up project a bit
We were running on an ancient version of Clojure, so that needed to be
updated, and while I was in the project.clj, I updated a lot of the
other dependencies. Finally, I wanted to simplify the code a touch, so I
did that with an `and` as opposed to an `if`, and things are running
just fine.