Added my own user for the future
We needed to have 'acl' installed for the 'setfacl -Rbn workspace'
command to remove all the advanced permissions that I don't want or
need. And then I uncommented the new user for the time when I want to
use the dotfiles and pull in my own account. If I get serious about
this, then I'm going to want to personalize it a lot, and make it
something I really use.